Added error handling

See merge request disi/innovation/aristote/aristoteapi!20

+++ src/Controller/Api/EnrichmentsController.php
@@ -4,6 +4,7 @@ namespace App\Controller\Api;
 
 use App\Entity\Enrichment;
 use App\Entity\EnrichmentVersion;
+use App\Model\ErrorsResponse;
 use App\Repository\EnrichmentRepository;
 use App\Repository\EnrichmentVersionRepository;
 use App\Service\ApiClientManager;
@@ -14,11 +15,13 @@ use OpenApi\Attributes as OA;
 use Psr\Log\LoggerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Bundle\SecurityBundle\Security;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Serializer\Normalizer\AbstractNormalizer;
 use Symfony\Component\Serializer\SerializerInterface;
+use Symfony\Component\Validator\Constraints\Uuid;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 #[Route('/v1')]
@@ -68,6 +71,14 @@ class EnrichmentsController extends AbstractController
             ]
         )
     )]
+    #[OA\Response(
+        response: 400,
+        description: 'Bad parameters',
+        content: new OA\JsonContent(
+            ref: new Model(type: ErrorsResponse::class),
+            type: 'object'
+        )
+    )]
     #[OA\Parameter(
         name: 'sort',
         description: 'Sort property to use',
@@ -98,6 +109,12 @@ class EnrichmentsController extends AbstractController
         $size = $request->query->get('size', 50);
         $page = $request->query->get('page', 1);
 
+        $paginationParametersErrors = $this->paginationUtils->paginationRequestParametersValidator(Enrichment::getSortFields(), $sort, $order, $size, $page);
+
+        if ([] !== $paginationParametersErrors) {
+            return $this->json(['status' => 'KO', 'errors' => $paginationParametersErrors], 400);
+        }
+
         $clientId = $this->security->getToken()->getAttribute('oauth_client_id');
         $clientEntity = $apiClientManager->getClientEntity($clientId);
 
@@ -123,6 +140,30 @@ class EnrichmentsController extends AbstractController
             ref: new Model(type: Enrichment::class, groups: ['enrichments'])
         )
     )]
+    #[OA\Response(
+        response: 400,
+        description: 'Bad parameters',
+        content: new OA\JsonContent(
+            ref: new Model(type: ErrorsResponse::class),
+            type: 'object'
+        )
+    )]
+    #[OA\Response(
+        response: 403,
+        description: 'Not allowed to access this resource',
+        content: new OA\JsonContent(
+            ref: new Model(type: ErrorsResponse::class),
+            type: 'object'
+        )
+    )]
+    #[OA\Response(
+        response: 404,
+        description: 'Entity not found',
+        content: new OA\JsonContent(
+            ref: new Model(type: ErrorsResponse::class),
+            type: 'object'
+        )
+    )]
     #[OA\Parameter(
         name: 'id',
         description: 'Enrichment ID',
@@ -132,7 +173,17 @@ class EnrichmentsController extends AbstractController
     #[Route('/enrichments/{id}', name: 'enrichment', methods: ['GET'])]
     public function getEnrichmentByID(string $id, ApiClientManager $apiClientManager, EnrichmentRepository $enrichmentRepository): Response
     {
+        $uuidValidationErrorResponse = $this->validateUuid($id);
+        if (null !== $uuidValidationErrorResponse) {
+            return $uuidValidationErrorResponse;
+        }
+
         $enrichment = $enrichmentRepository->findOneBy(['id' => $id]);
+
+        $enrichmentAccessErrorResponse = $this->validateObjectAccess($enrichment, $id);
+        if (null !== $enrichmentAccessErrorResponse) {
+            return $enrichmentAccessErrorResponse;
+        }
 
         $options = [
             AbstractNormalizer::GROUPS => ['enrichments'],
@@ -150,8 +201,54 @@ class EnrichmentsController extends AbstractController
         response: 200,
         description: 'Returns a list of enrichment versions',
         content: new OA\JsonContent(
-            type: 'array',
-            items: new OA\Items(ref: new Model(type: EnrichmentVersion::class, groups: ['enrichment_versions', 'enrichment_versions_with_transcript']))
+            type: 'object',
+            properties: [
+                new OA\Property(
+                    property: 'content',
+                    type: 'array',
+                    items: new OA\Items(ref: new Model(type: EnrichmentVersion::class, groups: ['enrichment_versions', 'enrichment_versions_with_transcript']))
+                ),
+                new OA\Property(
+                    property: 'isLastPage',
+                    description: 'Returns true if this is the last page, false otherwise',
+                    type: 'boolean'
+                ),
+                new OA\Property(
+                    property: 'currentPage',
+                    description: 'Returns the current page number',
+                    type: 'integer', format: 'int64'
+                ),
+
+                new OA\Property(
+                    property: 'totalElements',
+                    description: 'Returns the total number of elements',
+                    type: 'integer', format: 'int64'
+                ),
+            ]
+        )
+    )]
+    #[OA\Response(
+        response: 400,
+        description: 'Bad parameters',
+        content: new OA\JsonContent(
+            ref: new Model(type: ErrorsResponse::class),
+            type: 'object'
+        )
+    )]
+    #[OA\Response(
+        response: 403,
+        description: 'Not allowed to access this resource',
+        content: new OA\JsonContent(
+            ref: new Model(type: ErrorsResponse::class),
+            type: 'object'
+        )
+    )]
+    #[OA\Response(
+        response: 404,
+        description: 'Entity not found',
+        content: new OA\JsonContent(
+            ref: new Model(type: ErrorsResponse::class),
+            type: 'object'
         )
     )]
     #[OA\Parameter(
@@ -188,8 +285,18 @@ class EnrichmentsController extends AbstractController
         schema: new OA\Schema(type: 'boolean')
     )]
     #[Route('/enrichments/{id}/versions', name: 'enrichment_versions', methods: ['GET'])]
-    public function getEnrichmentVersionsByEnrichmentID(string $id, Request $request, ApiClientManager $apiClientManager, EnrichmentVersionRepository $enrichmentVersionRepository): Response
-    {
+    public function getEnrichmentVersionsByEnrichmentID(
+        string $id,
+        Request $request,
+        ApiClientManager $apiClientManager,
+        EnrichmentVersionRepository $enrichmentVersionRepository,
+        EnrichmentRepository $enrichmentRepository
+    ): Response {
+        $uuidValidationErrorResponse = $this->validateUuid($id);
+        if (null !== $uuidValidationErrorResponse) {
+            return $uuidValidationErrorResponse;
+        }
+
         $sort = $request->query->get('sort', 'createdAt');
         $order = $request->query->get('order', 'DESC');
         $size = $request->query->get('size', 2);
@@ -197,8 +304,21 @@ class EnrichmentsController extends AbstractController
         $withTranscript = $request->query->get('withTranscript', 'true');
         $groups = ['enrichment_versions'];
 
+        $paginationParametersErrors = $this->paginationUtils->paginationRequestParametersValidator(EnrichmentVersion::getSortFields(), $sort, $order, $size, $page);
+
+        if ([] !== $paginationParametersErrors) {
+            return $this->json(['status' => 'KO', 'errors' => $paginationParametersErrors], 400);
+        }
+
         if ('true' === $withTranscript) {
             $groups[] = 'enrichment_versions_with_transcript';
+        }
+
+        $enrichment = $enrichmentRepository->findOneBy(['id' => $id]);
+
+        $enrichmentAccessErrorResponse = $this->validateObjectAccess($enrichment, $id);
+        if (null !== $enrichmentAccessErrorResponse) {
+            return $enrichmentAccessErrorResponse;
         }
 
         $enrichmentVersions = $enrichmentVersionRepository->findByEnrichmentID($id, $page, $size, $sort, $order);
@@ -223,6 +343,30 @@ class EnrichmentsController extends AbstractController
             ref: new Model(type: EnrichmentVersion::class, groups: ['enrichment_versions', 'enrichment_versions_with_transcript'])
         )
     )]
+    #[OA\Response(
+        response: 400,
+        description: 'Bad parameters',
+        content: new OA\JsonContent(
+            ref: new Model(type: ErrorsResponse::class),
+            type: 'object'
+        )
+    )]
+    #[OA\Response(
+        response: 403,
+        description: 'Not allowed to access this resource',
+        content: new OA\JsonContent(
+            ref: new Model(type: ErrorsResponse::class),
+            type: 'object'
+        )
+    )]
+    #[OA\Response(
+        response: 404,
+        description: 'Entity not found',
+        content: new OA\JsonContent(
+            ref: new Model(type: ErrorsResponse::class),
+            type: 'object'
+        )
+    )]
     #[OA\Parameter(
         name: 'id',
         description: 'Enrichment ID',
@@ -230,9 +374,25 @@ class EnrichmentsController extends AbstractController
         schema: new OA\Schema(type: 'string')
     )]
     #[Route('/enrichments/{id}/versions/latest', name: 'latest_enrichment_version', methods: ['GET'])]
-    public function getLatestEnrichmentVersionByEnrichmentID(string $id, ApiClientManager $apiClientManager, EnrichmentVersionRepository $enrichmentVersionRepository): Response
+    public function getLatestEnrichmentVersionByEnrichmentID(string $id, ApiClientManager $apiClientManager, EnrichmentVersionRepository $enrichmentVersionRepository, EnrichmentRepository $enrichmentRepository): Response
     {
+        $uuidValidationErrorResponse = $this->validateUuid($id);
+        if (null !== $uuidValidationErrorResponse) {
+            return $uuidValidationErrorResponse;
+        }
+
+        $enrichment = $enrichmentRepository->findOneBy(['id' => $id]);
+
+        $enrichmentAccessErrorResponse = $this->validateObjectAccess($enrichment, $id);
+        if (null !== $enrichmentAccessErrorResponse) {
+            return $enrichmentAccessErrorResponse;
+        }
+
         $enrichmentVersions = $enrichmentVersionRepository->findBy(['enrichment' => $id], ['createdAt' => 'DESC']);
+
+        if ([] === $enrichmentVersions) {
+            return $this->json(['status' => 'KO', 'errors' => [sprintf("No version for enrichment with ID '%s' has been found", $id)]], 404);
+        }
 
         $options = [
             AbstractNormalizer::GROUPS => ['enrichment_versions', 'enrichment_versions_with_transcript'],
@@ -254,6 +414,30 @@ class EnrichmentsController extends AbstractController
             ref: new Model(type: EnrichmentVersion::class, groups: ['enrichment_versions', 'enrichment_versions_with_transcript'])
         )
     )]
+    #[OA\Response(
+        response: 400,
+        description: 'Bad parameters',
+        content: new OA\JsonContent(
+            ref: new Model(type: ErrorsResponse::class),
+            type: 'object'
+        )
+    )]
+    #[OA\Response(
+        response: 403,
+        description: 'Not allowed to access this resource',
+        content: new OA\JsonContent(
+            ref: new Model(type: ErrorsResponse::class),
+            type: 'object'
+        )
+    )]
+    #[OA\Response(
+        response: 404,
+        description: 'Entity not found',
+        content: new OA\JsonContent(
+            ref: new Model(type: ErrorsResponse::class),
+            type: 'object'
+        )
+    )]
     #[OA\Parameter(
         name: 'versionId',
         description: 'Enrichment version ID',
@@ -263,7 +447,17 @@ class EnrichmentsController extends AbstractController
     #[Route('/versions/{versionId}', name: 'enrichment_version', methods: ['GET'])]
     public function getEnrichmentVersionByID(string $versionId, ApiClientManager $apiClientManager, EnrichmentVersionRepository $enrichmentVersionRepository): Response
     {
+        $uuidValidationErrorResponse = $this->validateUuid($versionId);
+        if (null !== $uuidValidationErrorResponse) {
+            return $uuidValidationErrorResponse;
+        }
+
         $enrichmentVersion = $enrichmentVersionRepository->findOneBy(['id' => $versionId]);
+
+        $enrichmentVersionAccessErrorResponse = $this->validateObjectAccess($enrichmentVersion, $versionId, true);
+        if (null !== $enrichmentVersionAccessErrorResponse) {
+            return $enrichmentVersionAccessErrorResponse;
+        }
 
         $options = [
             AbstractNormalizer::GROUPS => ['enrichment_versions', 'enrichment_versions_with_transcript'],
@@ -291,20 +485,17 @@ class EnrichmentsController extends AbstractController
     )]
     #[OA\Response(
         response: 403,
-        description: 'Unsuccessful delete operation',
+        description: 'Delete operation not allowed',
         content: new OA\JsonContent(
-            properties: [
-                new OA\Property(
-                    property: 'status',
-                    description: 'KO',
-                    type: 'string'
-                ),
-                new OA\Property(
-                    property: 'error',
-                    description: 'Error message',
-                    type: 'string'
-                ),
-            ],
+            ref: new Model(type: ErrorsResponse::class),
+            type: 'object'
+        )
+    )]
+    #[OA\Response(
+        response: 404,
+        description: 'Entity not found',
+        content: new OA\JsonContent(
+            ref: new Model(type: ErrorsResponse::class),
             type: 'object'
         )
     )]
@@ -317,15 +508,52 @@ class EnrichmentsController extends AbstractController
     #[Route('/versions/{versionId}', name: 'delete_enrichment_version', methods: ['DELETE'])]
     public function deleteEnrichmentVersion(string $versionId, EnrichmentVersionRepository $enrichmentVersionRepository, EntityManagerInterface $entityManager): Response
     {
+        $uuidValidationErrorResponse = $this->validateUuid($versionId);
+        if (null !== $uuidValidationErrorResponse) {
+            return $uuidValidationErrorResponse;
+        }
+
         $enrichmentVersion = $enrichmentVersionRepository->findOneBy(['id' => $versionId]);
 
+        $enrichmentVersionAccessErrorResponse = $this->validateObjectAccess($enrichmentVersion, $versionId, true);
+        if (null !== $enrichmentVersionAccessErrorResponse) {
+            return $enrichmentVersionAccessErrorResponse;
+        }
+
         if ($enrichmentVersion->isInitialVersion()) {
-            return $this->json(['status' => 'KO', 'error' => "Can't delete initial version"], 403);
+            return $this->json(['status' => 'KO', 'errors' => "Can't delete initial version"], 403);
         }
 
         $entityManager->remove($enrichmentVersion);
         $entityManager->flush();
 
         return $this->json(['status' => 'OK']);
+    }
+
+    private function validateUuid(string $id): ?JsonResponse
+    {
+        $constraintViolationList = $this->validator->validate($id, new Uuid());
+
+        if ($constraintViolationList->count() > 0) {
+            return $this->json(['status' => 'KO', 'errors' => [sprintf("'%s' is not a valid UUID", $id)]], 400);
+        }
+
+        return null;
+    }
+
+    private function validateObjectAccess(Enrichment|EnrichmentVersion|null $object, string $id, bool $objectIntendedToBeEnrichmentVersion = false): ?JsonResponse
+    {
+        $objectName = sprintf('enrichment', $objectIntendedToBeEnrichmentVersion ? ' version' : '');
+        if (null === $object) {
+            return $this->json(['status' => 'KO', 'errors' => [sprintf("No %s with ID '%s' has been found", $objectName, $id)]], 404);
+        }
+
+        $enrichment = $object instanceof EnrichmentVersion ? $object->getEnrichment() : $object;
+
+        if ($enrichment->getCreatedBy()->getIdentifier() !== $this->security->getToken()->getAttribute('oauth_client_id')) {
+            return $this->json(['status' => 'KO', 'errors' => [sprintf('You are not allowed to access this %s', $objectName, $id)]], 403);
+        }
+
+        return null;
     }
 }
